+++ Neo/Gui/src/ThemedButton.cpp
@@ -129,8 +129,8 @@ void ThemedButton::draw()
 
 	Box3d* box = m_labelText->getBoundingBox();
 	//render->drawColoredQuad(m_x, m_y, m_width, m_height, color, m_rotation);
-	render->drawText(m_labelText, m_x + 0.5 * m_width,
-					 m_y + (box->max.y - box->min.y) + 0.5 * m_height,
+	render->drawText(m_labelText, m_x + 0.5 * (float) m_width,
+					 m_y + (box->max.y - box->min.y) + 0.5 * (float) m_height,
 					 m_rotation);
 }
